biopass_dao/requirements.txt

python-dotenv
psycopg2-binary
opencv-python
opencv-contrib-python
numpy
Pillow
